Move PrePurchaseNotices to composite-checkout directory
The `PrePurchaseNotices` component (and its dependencies) previously
lived in the old checkout directory, but since that directory will soon
be removed, this change moves it into the new checkout directory.
