+++ client/my-sites/checkout/composite-checkout/components/prepurchase-notices/prepurchase-notice.jsx
