+++ client/my-sites/checkout/composite-checkout/components/prepurchase-notices/style.scss
